Add Unit test & Intergation test
