+++ src/test/java/com/dawn/jat/illuminati/intergationtest/HelloControllerTest.java
@@ -1,0 +1,33 @@
+package com.dawn.jat.illuminati.intergationtest;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import com.dawn.jat.illuminati.hello.HelloController;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.client.TestRestTemplate;
+import org.springframework.boot.web.server.LocalServerPort;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class HelloControllerTest {
+    @Autowired
+    private HelloController controller;
+
+    @Test
+    void contextLoads() throws Exception {
+        assertThat(controller).isNotNull();
+    }
+
+    @LocalServerPort
+    private int port;
+
+    @Autowired
+    private TestRestTemplate restTemplate;
+
+    @Test
+    public void helloPageShouldReturnDefaultMessage() throws Exception {
+        assertThat(this.restTemplate.getForObject("http://localhost:" + port + "/hello",
+                String.class)).contains("Hello World!!!");
+    }
+}
